backend/requirements.txt
fastapi
uvicorn
joblib
numpy
pandas
xgboost
scikit-learn
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlalchemy
python-dotenv

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
